Move Options, Transients and Cache classes to sub-namespace Data_Storage
