+++ includes/class-wp-typography-factory.php
@@ -26,9 +26,9 @@
 
 use \Dice\Dice;
 
-use \WP_Typography\Cache;
-use \WP_Typography\Options;
-use \WP_Typography\Transients;
+use \WP_Typography\Data_Storage\Cache;
+use \WP_Typography\Data_Storage\Options;
+use \WP_Typography\Data_Storage\Transients;
 
 use \WP_Typography\Components\Admin_Interface;
 use \WP_Typography\Components\Multilingual;
