+++ includes/wp-typography/components/class-admin-interface.php
@@ -28,7 +28,7 @@
 
 namespace WP_Typography\Components;
 
-use \WP_Typography\Options;
+use \WP_Typography\Data_Storage\Options;
 use \WP_Typography\UI;
 use \WP_Typography\Settings\Plugin_Configuration as Config;
