+++ tests/data-storage/class-transients-test.php
@@ -22,9 +22,11 @@
  *  @license http://www.gnu.org/licenses/gpl-2.0.html
  */
 
-namespace WP_Typography\Tests;
+namespace WP_Typography\Tests\Data_Storage;
 
-use WP_Typography\Transients;
+use WP_Typography\Data_Storage\Transients;
+
+use WP_Typography\Tests\TestCase;
 
 use Brain\Monkey\Actions;
 use Brain\Monkey\Filters;
@@ -33,20 +35,20 @@ use Brain\Monkey\Functions;
 use Mockery as m;
 
 /**
- * WP_Typography\Transients unit test for the singleton methods.
+ * WP_Typography\Data_Storage\Transients unit test for the singleton methods.
  *
- * @coversDefaultClass \WP_Typography\Transients
- * @usesDefaultClass \WP_Typography\Transients
+ * @coversDefaultClass \WP_Typography\Data_Storage\Transients
+ * @usesDefaultClass \WP_Typography\Data_Storage\Transients
  *
  * @uses ::__construct
- * @uses \WP_Typography\Abstract_Cache::__construct
+ * @uses \WP_Typography\Data_Storage\Abstract_Cache::__construct
  */
 class Transients_Test extends TestCase {
 
 	/**
 	 * Test fixture.
 	 *
-	 * @var \WP_Typography\Transients
+	 * @var \WP_Typography\Data_Storage\Transients
 	 */
 	protected $transients;
 
@@ -74,7 +76,7 @@ class Transients_Test extends TestCase {
 	 *
 	 * @covers ::__construct
 	 *
-	 * @uses \WP_Typography\Abstract_Cache::__construct
+	 * @uses \WP_Typography\Data_Storage\Abstract_Cache::__construct
 	 */
 	public function test___construct() {
 		Functions\expect( 'get_transient' )->once()->with( Transients::INCREMENTOR_KEY )->andReturn( 0 );
